refactor(backend-game-application): extract GameEventV2FromGameActionBuilder extract builder from GameEventV2FromGameMessageEventBuilder
